Predict average slopes for several intervals using a model fitted by time_model().

Comoute average slopes for "clubic slope", "linear splines" and "cubic splines" fitted using time_model().

Usage

compute_slopes(
  fit,
  method,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = list(cubic_slope = NULL, linear_splines = c(5.5, 11), cubic_splines = c(1, 8,
    12))[[method]]
)

Arguments

fit

A model object from a statistical model such as from a call to time_model().

method

The type of model provided in fit, i.e., one of "cubic_slope", "linear_splines" or "cubic_splines".

period

The intervals knots on which slopes are to be computed.

knots

The knots as defined fit and according to method.

Value

A data.frame with slopes for each individuals/samples.

Examples

data("bmigrowth")
ls_mod <- time_model(
  x = "age",
  y = "log(bmi)",
  cov = NULL,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  method = "linear_splines"
)
#> nlme::lme(
#>   fixed = log(bmi) ~ gsp(age, knots = c(5.5, 11), degree = rep(1, 3), smooth = rep(0, 2)),
#>   data = data,
#>   random = ~ gsp(age, knots = c(5.5, 11), degree = rep(1, 3), smooth = rep(0, 2)) | ID,
#>   na.action = stats::na.omit,
#>   method = "ML",
#>   control = nlme::lmeControl(opt = "optim", maxIter = 500, msMaxIter = 500)
#> )
head(compute_slopes(
  fit = ls_mod,
  method = "linear_splines",
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17)#,
  # knots = list(
  #   "cubic_slope" = NULL,
  #   "linear_splines" = c(5.5, 11),
  #   "cubic_splines" = c(1, 8, 12)
  # )[[method]]
))
#>    ID pred_period_0 pred_period_0.5 pred_period_1.5 pred_period_3.5
#> 1 082      2.853184        2.886145        2.952068        3.083912
#> 2 083      2.738052        2.764628        2.817779        2.924081
#> 3 080      2.731110        2.759454        2.816140        2.929513
#> 4 031      2.673129        2.688156        2.718211        2.778319
#> 5 007      2.724079        2.776169        2.880347        3.088704
#> 6 033      2.768962        2.780169        2.802583        2.847412
#>   pred_period_6.5 pred_period_10 pred_period_12 pred_period_17 slope_0--0.5
#> 1        3.255880       3.396309       3.523904       3.961264   0.06592240
#> 2        3.094410       3.318502       3.450097       3.787938   0.05315120
#> 3        3.109982       3.344816       3.482937       3.838067   0.05668657
#> 4        2.922776       3.217996       3.340014       3.528365   0.03005422
#> 5        3.334483       3.465463       3.628122       4.254301   0.10417842
#> 6        2.970990       3.246609       3.358980       3.527090   0.02241449
#>   slope_1.5--3.5 slope_6.5--10 slope_12--17
#> 1     0.06592240    0.04012264   0.08747200
#> 2     0.05315120    0.06402628   0.06756830
#> 3     0.05668657    0.06709535   0.07102608
#> 4     0.03005422    0.08434848   0.03767016
#> 5     0.10417842    0.03742271   0.12523592
#> 6     0.02241449    0.07874839   0.03362208