Derived areas under the curve from a cubic splines mixed-effects model by egg_model().

Derived areas under the curve (AUCs) for differentintervals based on a fitted cubic splines mixed-effects model from egg_model(). This function is a specific version of compute_aucs designed to work specifically on egg_model().

Usage

egg_aucs(
  fit,
  period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
  knots = c(1, 8, 12)
)

Arguments

fit

A model object from a statistical model such as from a call to egg_model().

period

The intervals knots on which AUCs are to be computed.

knots

The knots as defined fit and according to method.

Value

A data.frame with AUC for each individuals/samples.

Examples

data("bmigrowth")
res <- egg_model(
  formula = log(bmi) ~ age,
  data = bmigrowth[bmigrowth[["sex"]] == 0, ],
  id_var = "ID",
  random_complexity = 1
)
#> Fitting model:
#>   nlme::lme(
#>     fixed = log(bmi) ~ gsp(age, knots = c(1, 8, 12), degree = rep(3, 4), smooth = rep(2, 3)),
#>     data = data,
#>     random = ~ gsp(age, knots = c(1, 8, 12), degree = rep(1, 4), smooth = rep(2, 3)) | ID,
#>     na.action = stats::na.omit,
#>     method = "ML",
#>     control = nlme::lmeControl(opt = "optim", niterEM = 25, maxIter = 500, msMaxIter = 500)
#>   )
head(
  egg_aucs(
    fit = res,
    period = c(0, 0.5, 1.5, 3.5, 6.5, 10, 12, 17),
    knots = c(1, 8, 12)
  )
)
#>      ID auc_0--0.5 auc_1.5--3.5 auc_6.5--10 auc_12--17
#> 082 082   1.385701     5.731551    11.21274   17.89646
#> 083 083   1.358239     5.539061    10.50629   16.31334
#> 080 080   1.362232     5.718337    11.55035   18.93891
#> 031 031   1.389514     5.708146    10.99889   17.32248
#> 007 007   1.389835     5.947658    12.48345   21.09766
#> 033 033   1.451842     6.026116    11.86239   19.03286